+++ controllers/user.go
@@ -3,10 +3,16 @@ package controllers
 import (
 	"ShareShelf/database/connection"
 	"ShareShelf/database/models"
+	"ShareShelf/helpers"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
+
+type TokenRequest struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
 
 func RegisterUser(context *gin.Context) {
 	var user models.User
@@ -27,4 +33,34 @@ func RegisterUser(context *gin.Context) {
 		return
 	}
 	context.JSON(http.StatusCreated, gin.H{"userId": user.ID, "email": user.Email, "username": user.Username})
+}
+
+func LoginUser(context *gin.Context) {
+	var request TokenRequest
+	var user models.User
+	if err := context.ShouldBindJSON(&request); err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		context.Abort()
+		return
+	}
+	// check if email exists and password is correct
+	record := connection.Instance.Where("email = ?", request.Email).First(&user)
+	if record.Error != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"error": record.Error.Error()})
+		context.Abort()
+		return
+	}
+	credentialError := user.CheckPassword(request.Password)
+	if credentialError != nil {
+		context.JSON(http.StatusUnauthorized, gin.H{"error": "invalid credentials"})
+		context.Abort()
+		return
+	}
+	tokenString, err := helpers.GenerateJWT(user.Email, user.Username)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		context.Abort()
+		return
+	}
+	context.JSON(http.StatusOK, gin.H{"token": tokenString})
 }
